@@ -1,12 +1,14 @@
 using System;
 using System.IO;
 using System.IO.Compression;
+using System.Linq;
 using System.Net;
 using System.Net.Http;
+using System.Net.Http.Headers;
 using System.Text;
-using Datadog.Trace.TestHelpers;
-using Newtonsoft.Json;
-using Directory = System.IO.Directory;
+using System.Text.Json;
+using System.Text.RegularExpressions;
+using System.Threading.Tasks;
 
 namespace UpdateVendors
 {
@@ -18,62 +20,66 @@ namespace UpdateVendors
 //------------------------------------------------------------------------------
 ";
 
-        private static readonly string CurrentDirectory = Environment.CurrentDirectory;
-        private static readonly string DownloadDirectory = Path.Combine(CurrentDirectory, "downloads");
+        private static readonly string DownloadDirectory = Path.Combine(Environment.CurrentDirectory, "downloads");
         private static string _vendorProjectDirectory;
 
-        public static void Main(string[] args)
+        public static async Task Main(string[] args)
         {
             InitializeCleanDirectory(DownloadDirectory);
-            var solutionDirectory = EnvironmentHelper.GetSolutionDirectory();
+            var solutionDirectory = GetSolutionDirectory();
             _vendorProjectDirectory = Path.Combine(solutionDirectory, "src", "Datadog.Trace", "Vendors");
             InitializeCleanDirectory(_vendorProjectDirectory);
 
-            UpdateVendor(
+            await UpdateVendorAsync(
                 libraryName: "Serilog",
                 masterBranchDownload: "https://github.com/serilog/serilog/archive/master.zip",
                 latestCommitUrl: "https://api.github.com/repos/serilog/serilog/commits/master",
                 pathToSrc: new[] { "serilog-master", "src", "Serilog" },
-                transform: TransformSerilog);
+                transform: filePath => RewriteCsFileWithStandardTransform(filePath, originalNamespace: "Serilog"));
 
-            UpdateVendor(
+            await UpdateVendorAsync(
                 libraryName: "Serilog.Sinks.File",
                 masterBranchDownload: "https://github.com/serilog/serilog-sinks-file/archive/master.zip",
                 latestCommitUrl: "https://api.github.com/repos/serilog/serilog-sinks-file/commits/master",
                 pathToSrc: new[] { "serilog-sinks-file-master", "src", "Serilog.Sinks.File" },
-                transform: TransformSerilog);
+                transform: filePath => RewriteCsFileWithStandardTransform(filePath, originalNamespace: "Serilog"));
+
+            await UpdateVendorAsync(
+                libraryName: "StatsdClient",
+                masterBranchDownload: "https://github.com/DataDog/dogstatsd-csharp-client/archive/3.3.0.zip",
+                latestCommitUrl: "https://api.github.com/repos/DataDog/dogstatsd-csharp-client/commits/3.3.0",
+                pathToSrc: new[] { "dogstatsd-csharp-client-3.3.0", "src", "StatsdClient" },
+                transform: filePath => RewriteCsFileWithStandardTransform(filePath, originalNamespace: "StatsdClient"));
         }
 
-        private static void TransformSerilog(string filePath)
+        private static void RewriteCsFileWithStandardTransform(string filePath, string originalNamespace)
         {
-            var extension = Path.GetExtension(filePath);
-
-            if (extension == ".cs")
+            if (string.Equals(Path.GetExtension(filePath), ".cs", StringComparison.OrdinalIgnoreCase))
             {
                 RewriteFileWithTransform(
                     filePath,
                     content =>
                     {
-                        // Prevent analyzer
-                        content = AutoGeneratedMessage + content;
+                        // Disable analyzer
+                        var builder = new StringBuilder(AutoGeneratedMessage, content.Length * 2);
+
+                        builder.Append(content);
 
                         // Prevent namespace conflicts
-                        content = content.Replace("using Serilog", "using Datadog.Trace.Vendors.Serilog");
-                        content = content.Replace("namespace Serilog", "namespace Datadog.Trace.Vendors.Serilog");
+                        builder.Replace($"using {originalNamespace}", $"using Datadog.Trace.Vendors.{originalNamespace}");
+                        builder.Replace($"namespace {originalNamespace}", $"namespace Datadog.Trace.Vendors.{originalNamespace}");
 
-                        // Expose nothing we don't intend to
-                        content = content.Replace("public class", "internal class");
-                        content = content.Replace("public static class", "internal static class");
-                        content = content.Replace("public abstract class", "internal abstract class");
-                        content = content.Replace("public sealed class", "internal sealed class");
-                        content = content.Replace("public interface", "internal interface");
-
-                        return content;
+                        // Don't expose anything we don't intend to
+                        // by replacing all "public" access modifiers with "internal"
+                        return Regex.Replace(
+                            builder.ToString(),
+                            @"public(\s+((abstract|sealed|static)\s+)?(partial\s+)?(class|struct|interface|enum|delegate))",
+                            match => $"internal{match.Groups[1]}");
                     });
             }
         }
 
-        private static void UpdateVendor(
+        private static async Task UpdateVendorAsync(
             string libraryName,
             string masterBranchDownload,
             string latestCommitUrl,
@@ -96,13 +102,7 @@ namespace UpdateVendors
 
             Console.WriteLine($"Unzipped {libraryName} upgrade.");
 
-            var sourceLocation = extractLocation;
-
-            foreach (var pathPart in pathToSrc)
-            {
-                sourceLocation = Path.Combine(sourceLocation, pathPart);
-            }
-
+            var sourceLocation = Path.Combine(pathToSrc.Prepend(extractLocation).ToArray());
             var projFile = Path.Combine(sourceLocation, $"{libraryName}.csproj");
 
             // Rename the proj file to a txt for reference
@@ -142,23 +142,22 @@ namespace UpdateVendors
             string commitInformation;
             using (var client = new HttpClient())
             {
-                client.DefaultRequestHeaders.UserAgent.Add(new System.Net.Http.Headers.ProductInfoHeaderValue("AppName", "1.0"));
-                client.DefaultRequestHeaders.Accept.Add(new System.Net.Http.Headers.MediaTypeWithQualityHeaderValue("application/json"));
-                client.DefaultRequestHeaders.Authorization = new System.Net.Http.Headers.AuthenticationHeaderValue("Token", githubToken);
-                var responseTask = client.GetStringAsync(latestCommitUrl);
-                responseTask.Wait();
-                commitInformation = FormatJson(responseTask.Result);
+                client.DefaultRequestHeaders.UserAgent.Add(new ProductInfoHeaderValue("Datadog.Trace.UpdateVendors", "1.0"));
+                client.DefaultRequestHeaders.Accept.Add(new MediaTypeWithQualityHeaderValue("application/json"));
+                client.DefaultRequestHeaders.Authorization = new AuthenticationHeaderValue("Token", githubToken);
+
+                commitInformation = await client.GetStringAsync(latestCommitUrl);
             }
 
+            // save commit information to a formatted json file
             var commitJsonPath = Path.Combine(sourceLocation, "commit-info.json");
-            File.WriteAllText(commitJsonPath, commitInformation);
+            SaveFormattedJson(commitInformation, commitJsonPath);
 
             // Move it all to the vendors directory
+            Console.WriteLine($"Copying source of {libraryName} to vendor project.");
             var vendorFinalPath = Path.Combine(_vendorProjectDirectory, libraryName);
             SafeDeleteDirectory(vendorFinalPath);
             Directory.Move(sourceLocation, vendorFinalPath);
-            Console.WriteLine($"Copying source of {libraryName} to vendor project.");
-
             Console.WriteLine($"Finished {libraryName} upgrade.");
         }
 
@@ -186,10 +185,42 @@ namespace UpdateVendors
             }
         }
 
-        private static string FormatJson(string json)
+        private static void SaveFormattedJson(string json, string path)
         {
-            dynamic parsedJson = JsonConvert.DeserializeObject(json);
-            return JsonConvert.SerializeObject(parsedJson, Formatting.Indented);
+            var options = new JsonWriterOptions { Indented = true };
+
+            using (var jsonDocument = JsonDocument.Parse(json))
+            using (var fileStream = File.OpenWrite(path))
+            using (var utf8JsonWriter = new Utf8JsonWriter(fileStream, options))
+            {
+                jsonDocument.WriteTo(utf8JsonWriter);
+            }
+        }
+
+        private static string GetSolutionDirectory()
+        {
+            var startDirectory = Environment.CurrentDirectory;
+            var currentDirectory = Directory.GetParent(startDirectory);
+            const string searchItem = @"Datadog.Trace.sln";
+
+            while (true)
+            {
+                var slnFile = currentDirectory.GetFiles(searchItem).SingleOrDefault();
+
+                if (slnFile != null)
+                {
+                    break;
+                }
+
+                currentDirectory = currentDirectory.Parent;
+
+                if (currentDirectory == null || !currentDirectory.Exists)
+                {
+                    throw new Exception($"Unable to find solution directory from: {startDirectory}");
+                }
+            }
+
+            return currentDirectory.FullName;
         }
     }
 }
